requirements.txt
Flask
googletrans
gTTS
opencv-python
pytesseract
PyPDF2
python-docx
sumy
urllib3
nltk
libgl1-mesa-glx